split build info inspector into it's own file
misc formatting for build deploy window
