+++ XRTK-Core/Packages/com.xrtk.core/Editor/BuildPipeline/BuildInfoInspector.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: de2d8bf1dd024eb2851f0fad06d3323b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 8ac5213854cf4dbabd140decf8df1946, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
